refactor: MainActivity ui 분리
